Moved Login to own function, ignored types in menu.
The login prompts were moved to their own function login().

The if-else block in main() was changed to a match-case block to clean
up the code and to restrict the user to numeric inputs only. Also added
a try block on the menu input to insult the user, if he tries to enter
anything other than a number.