cmdLogin now returns string value.
cmdLogin() now returns a string value, instead of None. If the string is
empty, it indicates a failed login. If the string is not empty, the
contents include the HTML from the response. That should be the HTML
home page of the user's default planet.

This commit does not include handling for said response, just returns it
as a string to the main function.