Fixed a few parms, and started new prompt-builder.
Fixed the parm in the sfc.py function to get welcome message. Soup
constructor now requires a string parm, but mine was still bytes.

Added new function in sfc.py to build the command prompt. This gets it
out of the main function and allows the prompt to be cleanly dynamic.

Added comments to cmdLogin.py indicating need for login validation.