Finished cmdLogin module. Added getUsername func.
The cmdLogin module is complete. It now checks for a successful login
and returns the player's home planet page.

Also added a getUsername() function to sfc module to independently get
the username from the planet home screeen. Still need to add logic that
will check if a proper username was returned.